+++ main.rb
@@ -1,3 +1,3 @@
-require_relative 'classes/person_class'
-require_relative 'classes/student_class'
-require_relative 'classes/teacher_class'
+require_relative 'classes/person'
+require_relative 'classes/student'
+require_relative 'classes/teacher'
